+++ test/PaySlipProblem.Tests/service/EmployeeDetailsReaderTest.cs
@@ -10,7 +10,7 @@ namespace PaySlipProblem.Tests.service
     {
         private const string FirstName = "James";
         private const string SurName = "Bond";
-        private const int AnnualSalary = 50000;
+        private const double AnnualSalary = 50000;
         private const double SuperRate = 9.5;
         private static readonly DateTime StartDate = new(2001, 3, 1);
         private static readonly DateTime EndDate = new(2001, 3, 31);
@@ -23,7 +23,7 @@ namespace PaySlipProblem.Tests.service
             
             consoleReader.Setup(c => c.ReadString("name")).Returns(FirstName);
             consoleReader.Setup(c => c.ReadString("surname")).Returns(SurName);
-            consoleReader.Setup(c => c.ReadInt("annual salary")).Returns(AnnualSalary);
+            consoleReader.Setup(c => c.ReadDouble("annual salary")).Returns(AnnualSalary);
             consoleReader.Setup(c => c.ReadDouble("super rate")).Returns(SuperRate);
             consoleReader.Setup(c => c.ReadDate("start date")).Returns(StartDate);
             consoleReader.Setup(c => c.ReadDate("end date")).Returns(EndDate);
